// github/types/paths/orgs/{org}/public_members/{username}.types.ts
// This code was automatically generated from an OpenAPI description.
// Do not edit this file. Edit the OpenAPI file instead.
// For more information, see https://github.com/pmcelhaney/counterfact/blob/main/docs/faq.md

import type { WideOperationArgument } from "../../../../../counterfact-types/index.ts";
import type { OmitValueWhenNever } from "../../../../../counterfact-types/index.ts";
import type { MaybePromise } from "../../../../../counterfact-types/index.ts";
import type { COUNTERFACT_RESPONSE } from "../../../../../counterfact-types/index.ts";
import type { Context } from "../../../../../routes/_.context.ts";
import type { ResponseBuilderFactory } from "../../../../../counterfact-types/index.ts";
import type { forbidden } from "../../../../#/components/responses/forbidden.js";

/**
 * Check if the provided user is a public member of the organization.
 */
export type orgsCheckPublicMembershipForUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: orgsCheckPublicMembershipForUser_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * The user can publicize their own membership. (A user cannot publicize the membership for another user.)
 *
 * Note that you'll need to set `Content-Length` to zero when calling out to this endpoint. For more information, see "[HTTP method](https://docs.github.com/rest/guides/getting-started-with-the-rest-api#http-method)."
 */
export type orgsSetPublicMembershipForAuthenticatedUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: orgsSetPublicMembershipForAuthenticatedUser_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      403: forbidden;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Removes the public membership for the authenticated user from the specified organization, unless public visibility is enforced by default.
 */
export type orgsRemovePublicMembershipForAuthenticatedUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: orgsRemovePublicMembershipForAuthenticatedUser_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type orgsCheckPublicMembershipForUser_Path = {
  org: string;
  username: string;
};

export type orgsSetPublicMembershipForAuthenticatedUser_Path = {
  org: string;
  username: string;
};

export type orgsRemovePublicMembershipForAuthenticatedUser_Path = {
  org: string;
  username: string;
};